Run clang-format over codebase (#2626)
* Update formatting

* Update format

* Update line numbers

+++ vowpalwabbit/example.cc
@@ -89,10 +89,10 @@ example& example::operator=(example&& other) noexcept
   test_only = other.test_only;
   end_pass = other.end_pass;
   sorted = other.sorted;
-VW_WARNING_STATE_PUSH
-VW_WARNING_DISABLE_DEPRECATED_USAGE
+  VW_WARNING_STATE_PUSH
+  VW_WARNING_DISABLE_DEPRECATED_USAGE
   in_use = other.in_use;
-VW_WARNING_STATE_POP
+  VW_WARNING_STATE_POP
 
   other.weight = 1.f;
 
@@ -113,24 +113,18 @@ VW_WARNING_STATE_POP
   other.test_only = false;
   other.end_pass = false;
   other.sorted = false;
-VW_WARNING_STATE_PUSH
-VW_WARNING_DISABLE_DEPRECATED_USAGE
+  VW_WARNING_STATE_PUSH
+  VW_WARNING_DISABLE_DEPRECATED_USAGE
   other.in_use = false;
-VW_WARNING_STATE_POP
+  VW_WARNING_STATE_POP
   return *this;
 }
 
 void example::delete_unions(void (*delete_label)(void*), void (*delete_prediction)(void*))
 {
-  if (delete_label)
-  {
-    delete_label(&l);
-  }
+  if (delete_label) { delete_label(&l); }
 
-  if (delete_prediction)
-  {
-    delete_prediction(&pred);
-  }
+  if (delete_prediction) { delete_prediction(&pred); }
 }
 
 float collision_cleanup(features& fs)
@@ -215,10 +209,8 @@ void copy_example_data(bool audit, example* dst, example* src, size_t label_size
 
 void move_feature_namespace(example* dst, example* src, namespace_index c)
 {
-  if (std::find(src->indices.begin(), src->indices.end(), c) == src->indices.end())
-    return;  // index not present in src
-  if (std::find(dst->indices.begin(), dst->indices.end(), c) == dst->indices.end())
-    dst->indices.push_back(c);
+  if (std::find(src->indices.begin(), src->indices.end(), c) == src->indices.end()) return;  // index not present in src
+  if (std::find(dst->indices.begin(), dst->indices.end(), c) == dst->indices.end()) dst->indices.push_back(c);
 
   auto& fdst = dst->feature_space[c];
   auto& fsrc = src->feature_space[c];
@@ -317,8 +309,7 @@ void free_flatten_example(flat_example* fec)
   if (fec)
   {
     fec->fs.~features();
-    if (fec->tag_len > 0)
-      free(fec->tag);
+    if (fec->tag_len > 0) free(fec->tag);
     free(fec);
   }
 }
@@ -383,7 +374,8 @@ std::string a_s_pred_to_string(const example& ec)
   std::stringstream strstream;
   strstream << "ec.pred.a_s[";
   for (uint32_t i = 0; i < ec.pred.a_s.size(); i++)
-  { strstream << "(" << i << " = " << ec.pred.a_s[i].action << ", " << ec.pred.a_s[i].score << ")"; } strstream << "]";
+  { strstream << "(" << i << " = " << ec.pred.a_s[i].action << ", " << ec.pred.a_s[i].score << ")"; }
+  strstream << "]";
   return strstream.str();
 }
 
@@ -412,8 +404,7 @@ namespace VW
 example* alloc_examples(size_t, size_t count = 1)
 {
   example* ec = calloc_or_throw<example>(count);
-  if (ec == nullptr)
-    return nullptr;
+  if (ec == nullptr) return nullptr;
   for (size_t i = 0; i < count; i++)
   {
     ec[i].ft_offset = 0;
@@ -439,10 +430,7 @@ void finish_example(vw& all, multi_ex& ec_seq)
 
 void return_multiple_example(vw& all, v_array<example*>& examples)
 {
-  for (auto ec : examples)
-  {
-    clean_example(all, *ec, true);
-  }
+  for (auto ec : examples) { clean_example(all, *ec, true); }
   examples.clear();
 }
